Moved Round and Background Picture to their own files
