+++ backend/app.py
@@ -104,26 +104,6 @@ hardleaderboard_ref = db.collection('HardLeaderboard')
 players_ref = db.collection('players')
 
 
-class BackgroundPicture:
-    def _init_(self, document):
-        self.targetImage = ''  # TODO
-        self.filepath = document['url']
-        self.targetDate = document['date']
-        # TL, TR, BL, BR (for background currently but idk if we need these. Should use for target)
-        self.imageCoordinates = [document['BL'],
-                                 document['TL'], document['TR'], document['BR']]
-
-
-class Round:
-    # id from array in easy/medium/hard game collection, difficulty -> EasyRounds, MediumRounds, HardRounds
-    def _init_(self, ID, difficulty):
-        round_ref = db.collection(difficulty).document(ID)
-        self.averagePerformance = 0
-        self.numPlayersCompleted = 0
-        self.backgroundImage = BackgroundPicture(round_ref.get())
-        self.targetImage = ''  # TODO
-
-
 def get_randImage():  # gets random image from database
     docs = images_ref.get()
     documents_list = [doc.to_dict() for doc in docs]
